UPDATE: notification service created

+++ Controllers/ProjectController.cs
@@ -20,10 +20,14 @@ namespace ProjectManagementService.Controllers
             _projectService = projectService;
         }
 
-        private bool ValidateCurrentUser(string userId)
+        private async Task<bool> UserHasAccessToProject(string projectId)
         {
-            var currentUserId = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
-            return currentUserId == userId;
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (string.IsNullOrEmpty(userId))
+            {
+                return false;
+            }
+            return await _projectService.UserHasAccessToProject(userId, projectId);
         }
 
         [HttpGet]
@@ -58,138 +62,97 @@ namespace ProjectManagementService.Controllers
         [Authorize]
         public async Task<ActionResult<ProjectDTO>> CreateProject(CreateProjectRequest project)
         {
-            try
+            if (!ModelState.IsValid)
             {
-                // Initialize a list to store the task IDs
-                var taskIds = new List<string>();
+                return BadRequest(ModelState);
+            }
 
-                // Check if tasks are provided in the request
-                if (project.Tasks != null && project.Tasks.Count > 0)
+            var taskIds = new List<string>();
+
+            if (project.Tasks != null && project.Tasks.Any())
+            {
+                foreach (var taskRequest in project.Tasks)
                 {
-                    // Create each task and collect its ID
-                    foreach (var taskRequest in project.Tasks)
+                    if (!Enum.TryParse<ProjectTaskStatus>(taskRequest.Status, true, out var taskStatus))
                     {
-                        if (!Enum.TryParse<ProjectTaskStatus>(taskRequest.Status, true, out var taskStatus))
-                        {
-                            return BadRequest("Invalid task status value.");
-                        }
-
-                        var newTask = new ProjectTask
-                        {
-                            Title = taskRequest.Title,
-                            Description = taskRequest.Description,
-                            AssignedTo = taskRequest.AssignedTo,
-                            DueDate = taskRequest.DueDate,
-                            Status = taskStatus
-                        };
-
-                        // Create the task in the database
-                        var taskId = await _projectService.CreateTask(newTask);
-
-                        // Check if the task was created successfully
-                        if (taskId == null)
-                        {
-                            return StatusCode(500, "Failed to create one of the tasks.");
-                        }
-
-                        // Add the created task ID to the list
-                        taskIds.Add(taskId);
+                        ModelState.AddModelError($"Tasks[{project.Tasks.IndexOf(taskRequest)}].Status", "Invalid task status value.");
+                        return BadRequest(ModelState);
                     }
+
+                    var newTask = new ProjectTask
+                    {
+                        Title = taskRequest.Title,
+                        Description = taskRequest.Description,
+                        AssignedTo = taskRequest.AssignedTo,
+                        DueDate = taskRequest.DueDate,
+                        Status = taskStatus
+                    };
+
+                    var taskId = await _projectService.CreateTask(newTask);
+                    if (string.IsNullOrEmpty(taskId))
+                    {
+                        return StatusCode(500, "Failed to create one of the tasks.");
+                    }
+
+                    taskIds.Add(taskId);
                 }
-
-                // Create the new project with the collected task IDs
-                var newProject = new Project
-                {
-                    Name = project.Name,
-                    Description = project.Description,
-                    StartDate = project.StartDate,
-                    EndDate = project.EndDate,
-                    ProjectMembers = project.ProjectMemberIds,
-                    ProjectOwner = project.ProjectOwner,
-                    TaskIds = taskIds,
-                    Status = ProjectStatus.NotStarted
-                };
-
-                // Create the project in the database
-                var createdProject = await _projectService.CreateProject(newProject);
-
-                var projectDTO = new ProjectDTO
-                {
-                    Id = createdProject.Id,
-                    Name = createdProject.Name,
-                    Description = createdProject.Description,
-                    StartDate = createdProject.StartDate,
-                    EndDate = createdProject.EndDate,
-                    ProjectMembers = createdProject.ProjectMembers,
-                    ProjectOwner = createdProject.ProjectOwner,
-                    TaskIds = createdProject.TaskIds,
-                    Status = createdProject.Status.ToString()
-                };
-
-                // Check if the project was created successfully
-                if (createdProject == null)
-                {
-                    return StatusCode(500, "Failed to create the project.");
-                }
-
-                // Return the created project details
-                return CreatedAtRoute("GetProject", new { id = projectDTO.Id }, projectDTO);
             }
-            catch (Exception e)
+
+            var newProject = new Project
             {
-                // Log the exception (optional) and return a 500 status code with error message
-                Console.WriteLine($"Error creating project: {e.Message}");
-                return StatusCode(500, e.Message);
+                Name = project.Name,
+                Description = project.Description,
+                StartDate = project.StartDate,
+                EndDate = project.EndDate,
+                ProjectMembers = project.ProjectMemberIds,
+                ProjectOwner = project.ProjectOwner,
+                TaskIds = taskIds,
+                Status = ProjectStatus.NotStarted
+            };
+
+            var createdProject = await _projectService.CreateProject(newProject);
+            if (createdProject == null)
+            {
+                return StatusCode(500, "Failed to create the project.");
             }
+
+            var projectDTO = MapProjectToDTO(createdProject);
+
+            return CreatedAtRoute("GetProject", new { id = projectDTO.Id }, projectDTO);
+        }
+
+        private static ProjectDTO MapProjectToDTO(Project project)
+        {
+            return new ProjectDTO
+            {
+                Id = project.Id,
+                Name = project.Name,
+                Description = project.Description,
+                StartDate = project.StartDate,
+                EndDate = project.EndDate,
+                ProjectMembers = project.ProjectMembers,
+                ProjectOwner = project.ProjectOwner,
+                TaskIds = project.TaskIds,
+                Status = project.Status.ToString()
+            };
         }
 
         [HttpGet("{id:length(24)}", Name = "GetProject")]
         [Authorize]
         public async Task<ActionResult<ProjectDTO>> GetProjectById(string id)
         {
-            try
+            if (!await UserHasAccessToProject(id))
             {
-                var userId = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
-                if (string.IsNullOrEmpty(userId))
-                {
-                    return Unauthorized("User ID claim is missing or invalid.");
-                }
-
-                var project = await _projectService.GetProjectById(id);
-                if (project == null)
-                {
-                    return NotFound();
-                }
-
-                if (string.IsNullOrEmpty(project.Id))
-                {
-                    return BadRequest("Project ID is missing or invalid.");
-                }
-
-                if (!await _projectService.UserHasAccessToProject(userId, project.Id))
-                {
-                    return Unauthorized();
-                }
-
-                var projectDTO = new ProjectDTO
-                {
-                    Id = project.Id,
-                    Name = project.Name,
-                    Description = project.Description,
-                    StartDate = project.StartDate,
-                    EndDate = project.EndDate,
-                    ProjectMembers = project.ProjectMembers,
-                    ProjectOwner = project.ProjectOwner,
-                    TaskIds = project.TaskIds,
-                    Status = project.Status.ToString()
-                };
-
-                return Ok(projectDTO);
+                return Forbid();
             }
-            catch (Exception e)
+
+            var project = await _projectService.GetProjectById(id);
+            if (project == null)
             {
-                return StatusCode(500, e.Message);
+                return NotFound();
             }
+
+            return Ok(MapProjectToDTO(project));
         }
 
         [HttpPut("{id:length(24)}")]
@@ -198,16 +161,18 @@ namespace ProjectManagementService.Controllers
         {
             try
             {
-                // Fetch the project by ID
+                if (!await UserHasAccessToProject(id))
+                {
+                    return Unauthorized();
+                }
+
                 var projectIn = await _projectService.GetProjectById(id);
 
-                // Check if the project exists
                 if (projectIn == null)
                 {
                     return NotFound($"Project with ID {id} not found.");
                 }
 
-                // Update project properties from the request
                 projectIn.Name = project.Name;
                 projectIn.Description = project.Description;
                 projectIn.StartDate = project.StartDate;
@@ -215,7 +180,6 @@ namespace ProjectManagementService.Controllers
                 projectIn.ProjectMembers = project.ProjectMemberIds;
                 projectIn.ProjectOwner = project.ProjectOwner;
 
-                // Update the project in the database
                 var updatedProject = await _projectService.UpdateProject(id, projectIn);
 
                 if (updatedProject == null)
@@ -223,17 +187,14 @@ namespace ProjectManagementService.Controllers
                     return StatusCode(500, "Failed to update the project.");
                 }
 
-                // Return the updated project
                 return Ok(updatedProject);
             }
             catch (Exception e)
             {
-                // Log the exception (optional)
                 Console.WriteLine($"Error updating project: {e.Message}");
                 return StatusCode(500, "An error occurred while updating the project.");
             }
         }
-
 
         [HttpDelete("{id:length(24)}")]
         [Authorize]
@@ -241,6 +202,11 @@ namespace ProjectManagementService.Controllers
         {
             try
             {
+                if (!await UserHasAccessToProject(id))
+                {
+                    return Unauthorized();
+                }
+
                 var project = await _projectService.GetProjectById(id);
 
                 if (project == null)
@@ -264,6 +230,11 @@ namespace ProjectManagementService.Controllers
         {
             try
             {
+                if (!await UserHasAccessToProject(id))
+                {
+                    return Unauthorized();
+                }
+
                 await _projectService.AddProjectMember(id, userId);
                 return NoContent();
             }
@@ -279,6 +250,11 @@ namespace ProjectManagementService.Controllers
         {
             try
             {
+                if (!await UserHasAccessToProject(id))
+                {
+                    return Unauthorized();
+                }
+
                 await _projectService.RemoveProjectMember(id, userId);
                 return NoContent();
             }
@@ -294,13 +270,16 @@ namespace ProjectManagementService.Controllers
         {
             try
             {
-                // Validate and convert task status from string to enum
+                if (!await UserHasAccessToProject(id))
+                {
+                    return Unauthorized();
+                }
+
                 if (!Enum.TryParse<ProjectTaskStatus>(task.Status, true, out var taskStatus))
                 {
                     return BadRequest("Invalid status value.");
                 }
 
-                // Create new task object
                 var newTask = new ProjectTask
                 {
                     Title = task.Title,
@@ -310,33 +289,27 @@ namespace ProjectManagementService.Controllers
                     Status = taskStatus
                 };
 
-                // Create task in database and get its ID
                 var taskId = await _projectService.CreateTask(newTask);
 
-                // Check if task ID was generated correctly
                 if (taskId == null)
                 {
                     return StatusCode(500, "Task ID was not generated correctly.");
                 }
 
-                // Add task ID to the project
                 bool taskAdded = await _projectService.AddTask(id, taskId);
                 if (!taskAdded)
                 {
                     return StatusCode(500, "Failed to add task to the project.");
                 }
 
-                // Return the created task details
                 return CreatedAtRoute("GetTask", new { id = taskId }, newTask);
             }
             catch (Exception e)
             {
-                // Log the exception (optional) and return a 500 status code with error message
                 Console.WriteLine($"Error adding task: {e.Message}");
                 return StatusCode(500, "An error occurred while adding the task.");
             }
         }
-
 
         [HttpPut("{id:length(24)}/tasks/{taskId:length(24)}")]
         [Authorize]
@@ -344,6 +317,11 @@ namespace ProjectManagementService.Controllers
         {
             try
             {
+                if (!await UserHasAccessToProject(id))
+                {
+                    return Unauthorized();
+                }
+
                 if (!Enum.TryParse<ProjectTaskStatus>(task.Status, true, out var taskStatus))
                 {
                     return BadRequest("Invalid status value.");
@@ -372,6 +350,11 @@ namespace ProjectManagementService.Controllers
         {
             try
             {
+                if (!await UserHasAccessToProject(id))
+                {
+                    return Unauthorized();
+                }
+
                 await _projectService.RemoveTask(id, taskId);
                 return NoContent();
             }
